Move more helper functions to helpers.py
